+++ src/frontend/src/functions/forms.tsx
@@ -1,6 +1,6 @@
+import { invalidResponse, permissionDenied } from '@lib/functions/Notification';
 import type { ApiFormFieldSet, ApiFormFieldType } from '@lib/types/Forms';
 import type { AxiosResponse } from 'axios';
-import { invalidResponse, permissionDenied } from './notifications';
 
 /**
  * Extract the available fields (for a given method) from the response object
